+++ CHANGELOG.md
@@ -36,6 +36,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed issue with CMAccounts with Get returning a string when only one account exists.
 - Fixed issue with Boundary groups containing VPN type boundaries.
 
+### Changed
+
+- Added support for VPN and IPv6Prefix boundaries to the CMBoundaries resource.
+
 ## [1.0.2] - 2021-05-12
 
 ### Added